Rename AugmentedSentenceSequenceTagger to PrefixedSequenceTagger and move to own package
